pyproject.toml
[project]
name = "ai-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.0",
    "requests==2.32.3",
    "streamlit==1.40.0",
]
